requirements.txt
langchain
langchain-core
langchain-community
langgraph
langchain-google-genai
langchain-qdrant
qdrant-client
langchain-groq
python-dotenv